Adding filtration over source channel(s) for individual reads via new `--channels` flag.
When not specified, channels source (or lack thereof) is not considered.

`--channels` takes channel ids as `a(-b)`` integer(ranges) where `a` and `b` are positive integers, and `b` has to be greater or equal than `a`. Multiple entries can be listed, ranges overlaps are allowed, and the following entries are equivalent:
```
--channels 100 110-115 200
--channels 110-115 200 100
```
resulting in reads from channels `100`, `110`, `111`, `112`, `113`, `114`, `115`, `200` being retained, while other reads skipped.

Channel data for a fastq read is parsed from its header comment section with the first `ch=\d+` entry being used.